pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.6.3",
    "langchain-community>=0.3.19",
    "langchain>=0.3.20",
    "llama-index>=0.12.22",
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.0",
    "scikit-learn>=1.6.1",
    "numpy>=2.2.3",
    "langchain-openai>=0.3.9",
]
